Move viewer testing code to testing file
