+++ tests/cypress/e2e/datasets/imagegallery.cy.js
@@ -1,4 +1,4 @@
-import { retryableBefore, stringToArray } from '../../support/utils.js'
+import { retryableBefore, stringToArray, randomInteger } from '../../support/utils.js'
 
 /**
  * List of dataset ids
@@ -81,7 +81,99 @@ datasetIds.forEach((datasetId) => {
                 this.skip()
               }
             })
-            cy.checkGalleryItemViewer(datasetId, item)
+            cy.get('.el-cascader-node').contains(new RegExp(item, 'i')).parent().siblings().click()
+            cy.window().then((window) => {
+              cy.stub(window.document.body, 'appendChild').as('cardClicked')
+            })
+            cy.get('.el-pager > .number').then(($pages) => {
+              if ($pages.length > 1) {
+                const randomPage = randomInteger(0, $pages.length - 1)
+                cy.wrap($pages).eq(randomPage).click()
+              }
+              cy.get('.el-card > .el-card__body').then(($cards) => {
+                const randomCard = randomInteger(0, $cards.length - 1)
+                cy.wrap($cards).eq(randomCard).then(($card) => {
+                  cy.wrap($card).find('.el-button').click()
+                  cy.get('@cardClicked').should('be.calledWith', Cypress.sinon.match.any).then((stub) => {
+                    // Appended link element
+                    const element = stub.args[0][0]
+                    expect(element, 'Button should open a new tab').to.have.attr('target').to.contain('blank')
+                    cy.visit(element.href)
+                    cy.waitForPageLoading()
+                    if (item === 'Scaffold' || item === 'Flatmap') {
+                      // Check whether map viewer loaded or not
+                      cy.waitForViewerContainer('.mapClass')
+                      cy.get('.page-hero > .container').should(($content) => {
+                        expect($content, 'Map Viewer should display content').to.contain('Maps')
+                      })
+                      if (item === 'Scaffold') {
+                        cy.get('.pane-1 > .content-container > .toolbar > .toolbar-flex-container > .el-select > .el-select__wrapper > .el-select__selection > .el-select__placeholder > span').should(($title) => {
+                          expect($title, 'Map Viewer should display scaffold').to.contain('Scaffold')
+                        })
+                      }
+                      if (item === 'Flatmap') {
+                        cy.get('.toolbar-title').should(($title) => {
+                          expect($title, 'Map Viewer should display flatmap').to.contain('MultiFlatmap')
+                        })
+                      }
+                      cy.backToDetailPage(datasetId)
+                      cy.waitForPageLoading()
+                    } else {
+                      // Check whether metadata exist or not
+                      cy.waitForViewerContainer('.subpage')
+                      if (item === 'Video') {
+                        cy.get('video')
+                          .should('have.prop', 'paused', true)
+                          .and('have.prop', 'ended', false)
+                          .then(($video) => {
+                            $video[0].play()
+                          })
+                        // once the video starts playing, check props
+                        cy.get('video')
+                          .should('have.prop', 'paused', false)
+                          .and('have.prop', 'ended', false)
+                      }
+                      if (item === 'Segmentation') {
+                        cy.get('.biolucida-viewer > p > a').then(($link) => {
+                          expect($link, 'Button should open a new tab').to.have.attr('target').to.contain('blank')
+                        })
+                      }
+                      if (item === 'Image') {
+                        cy.window().then((window) => {
+                          cy.stub(window, 'open').as('Open')
+                        })
+                        cy.get('.biolucida-viewer > .el-row > div > .el-button').each(($button) => {
+                          cy.wrap($button).click()
+                          cy.get('@Open').should('be.calledWith', Cypress.sinon.match.any).then((stub) => {
+                            const link = stub.args[0][0]
+                            expect(link.length, 'Button should contain link to external resource').to.be.greaterThan(0)
+                          })
+                        })
+                      }
+                      cy.get('.subpage > .file-detail > :nth-child(2)').each(($row) => {
+                        expect($row.text().length, 'Viewer metadata should exist').to.be.greaterThan(0)
+                      })
+                      cy.get('.subpage').contains('File location').siblings().find('a').then(($link) => {
+                        cy.wrap($link).click()
+                        cy.waitForPageLoading()
+                        const breadcrumbs = $link.text().split('/')
+                        cy.get('.breadcrumb-list .breadcrumb-link').each(($breadcrumb) => {
+                          expect(breadcrumbs, 'Filepath should contain all breadcrumbs').to.include($breadcrumb.text())
+                        })
+                        const filenameList = breadcrumbs[breadcrumbs.length - 1].match(/[a-z0-9]+/gi)
+                        const regex = new RegExp(filenameList.join('.*'), 'gi')
+                        cy.get('.truncated > a').then(($filenames) => {
+                          expect($filenames.text(), 'Filename should exist in the table').to.match(regex)
+                        })
+                      })
+                    }
+                    cy.then(() => {
+                      cy.clickOnDetailTab('Gallery')
+                    })
+                  })
+                })
+              })
+            })
           } else {
             this.skip()
           }
